requirements/docs.txt
attrs==21.4.0
    # via
    #   -c D:\dgpy-dev\dgpy-libs\requirements\dev.txt
    #   jsonschema
beautifulsoup4==4.11.1
    # via nbconvert
bleach==5.0.1
    # via nbconvert
click==8.1.3
    # via
    #   -c D:\dgpy-dev\dgpy-libs\requirements\dev.txt
    #   mkdocs
colorama==0.4.6
    # via
    #   -c D:\dgpy-dev\dgpy-libs\requirements\dev.txt
    #   click
defusedxml==0.7.1
    # via nbconvert
entrypoints==0.4
    # via
    #   jupyter-client
    #   nbconvert
fastjsonschema==2.16.1
    # via nbformat
ghp-import==2.1.0
    # via mkdocs
griffe==0.22.0
    # via mkdocstrings-python
importlib-metadata==4.12.0
    # via mkdocs
jinja2==3.1.2
    # via
    #   mkdocs
    #   mkdocs-material
    #   mkdocstrings
    #   nbconvert
jsonschema==4.7.2
    # via nbformat
jupyter-client==7.3.4
    # via nbclient
jupyter-core==4.11.1
    # via
    #   jupyter-client
    #   nbconvert
    #   nbformat
jupyterlab-pygments==0.2.2
    # via nbconvert
jupytext==1.14.0
    # via mkdocs-jupyter
markdown==3.3.7
    # via
    #   markdown-callouts
    #   markdown-include
    #   mkdocs
    #   mkdocs-autorefs
    #   mkdocs-material
    #   mkdocstrings
    #   pymdown-extensions
markdown-callouts==0.3.0
    # via -r D:\dgpy-dev\dgpy-libs\requirements\docs.in
markdown-include==0.8.1
    # via -r D:\dgpy-dev\dgpy-libs\requirements\docs.in
markdown-it-py==2.2.0
    # via
    #   jupytext
    #   mdit-py-plugins
markupsafe==2.1.1
    # via
    #   jinja2
    #   mkdocstrings
    #   nbconvert
mdit-py-plugins==0.3.0
    # via jupytext
mdurl==0.1.1
    # via markdown-it-py
mergedeep==1.3.4
    # via mkdocs
mistune==0.8.4
    # via nbconvert
mkdocs==1.3.1
    # via
    #   -r D:\dgpy-dev\dgpy-libs\requirements\docs.in
    #   mkdocs-autorefs
    #   mkdocs-jupyter
    #   mkdocs-material
    #   mkdocstrings
mkdocs-autorefs==0.4.1
    # via mkdocstrings
mkdocs-jupyter==0.21.0
    # via -r D:\dgpy-dev\dgpy-libs\requirements\docs.in
mkdocs-material==8.3.9
    # via
    #   -r D:\dgpy-dev\dgpy-libs\requirements\docs.in
    #   mkdocs-jupyter
mkdocs-material-extensions==1.0.3
    # via mkdocs-material
mkdocstrings[python]==0.19.0
    # via
    #   -r D:\dgpy-dev\dgpy-libs\requirements\docs.in
    #   mkdocstrings-python
mkdocstrings-python==0.7.1
    # via mkdocstrings
nbclient==0.6.6
    # via nbconvert
nbconvert==6.5.0
    # via mkdocs-jupyter
nbformat==5.4.0
    # via
    #   jupytext
    #   nbclient
    #   nbconvert
nest-asyncio==1.5.5
    # via
    #   jupyter-client
    #   nbclient
packaging==21.3
    # via
    #   -c D:\dgpy-dev\dgpy-libs\requirements\dev.txt
    #   mkdocs
    #   nbconvert
pandocfilters==1.5.0
    # via nbconvert
pygments==2.12.0
    # via
    #   mkdocs-jupyter
    #   mkdocs-material
    #   nbconvert
pymdown-extensions==9.5
    # via
    #   mkdocs-material
    #   mkdocstrings
pyparsing==3.0.9
    # via
    #   -c D:\dgpy-dev\dgpy-libs\requirements\dev.txt
    #   packaging
pyrsistent==0.18.1
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   ghp-import
    #   jupyter-client
pywin32==306
    # via jupyter-core
pyyaml==6.0
    # via
    #   jupytext
    #   mkdocs
    #   pyyaml-env-tag
pyyaml-env-tag==0.1
    # via mkdocs
pyzmq==23.2.0
    # via jupyter-client
six==1.16.0
    # via
    #   -c D:\dgpy-dev\dgpy-libs\requirements\dev.txt
    #   bleach
    #   python-dateutil
soupsieve==2.3.2.post1
    # via beautifulsoup4
tinycss2==1.1.1
    # via nbconvert
toml==0.10.2
    # via
    #   -c D:\dgpy-dev\dgpy-libs\requirements\dev.txt
    #   jupytext
tornado==6.2
    # via jupyter-client
traitlets==5.3.0
    # via
    #   jupyter-client
    #   jupyter-core
    #   nbclient
    #   nbconvert
    #   nbformat
watchdog==2.1.9
    # via mkdocs
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
zipp==3.8.1
    # via importlib-metadata
